Fixed weird loop bug!
I had to remove the while loops from the client so that it only reads back a single line from the server on each request. I think all of the server commands return one-line outputs anyway, so this should be fine. Just in case, I modified the server so that it'll combine the command output into one line before sending it.